✨ 꼬깃 보관함 메인과 상세페이지 연결
